removed warnings, remarks and minor bugs
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@5438 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/vxdCaTracking/src/VXDTFTrackCandidate.cc
@@ -169,7 +169,7 @@ void VXDTFTrackCandidate::setCondition(bool newCondition)
 }
 
 
-void VXDTFTrackCandidate::setNeuronValue(float aValue) { m_neuronValue = aValue; }
+void VXDTFTrackCandidate::setNeuronValue(double aValue) { m_neuronValue = aValue; }
 
 void VXDTFTrackCandidate::removeVirtualHit()   /// removing virtual hit/segment from TC after collecting the whole TC. This is done after the TCC and the TCC filter since the TCC needs the information provided by the virtual hit/segment.
 {
